Remove unnecessary code and add funtionality to (-) button
When pressing the equals button, instead of having an operand2,
it is better to just use what is currently being displayed.

As such, code that was used to track the second operand
and then use it is no longer needed.

The (-) button will function properly as a negative button.
It will delete the current display and replace it with -0 if
selected first.

If the button is pressed after numbers are in the display it
will simply put a negative sign in front of the current number.